Restructuring web routes
