+++ app/templates/users.html
@@ -69,7 +69,7 @@ Users
 
 		{% else %}
 			<h3>User List
-				{% if current_user.role == 'admin' %}
+				{% if current_user.role.name == 'ADMIN' %}
 					<button type="button" class="btn btn-default btn-xs pull-right" data-toggle="modal" data-target="#addUserModal">Add User</button>
 				{% endif %}
 			</h3>
