+++ docker-compose.yml
@@ -14,10 +14,10 @@ services:
   app:
     build: ./
     environment:
-      - FLASK_APP=app.app
+      - FLASK_APP=app:app
       - FLASK_DEBUG=1
       - BASE_DIR=/opt/flask-app
-      - SQLALCHEMY_DATABASE_URI="postgresql://postgres:postgrespswd@db:5432/cloud_pcap"
+      - SQLALCHEMY_DATABASE_URI=postgresql://postgres:postgrespswd@db:5432/cloud_pcap
     depends_on:
       - db
     ports:
